backend/requirements.txt
# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# HTTP Client
httpx>=0.27.0

# AI / LLM
anthropic>=0.40.0

# Excel Export
openpyxl>=3.1.0

# HTML Parsing (SEC 8-K tracker)
beautifulsoup4>=4.12.0

# Utilities
python-dateutil>=2.8.0
